Register page basically done
Register page, writing to database working, passwords are hashed using the password_hash function so plan text isnt saved in the datebase. Additional a basic structure for the sign in page is done. Need to read from database to find email then use
password_verify building function to test equality between the user's entered plaintext and the hashed value in the database